requirements.txt
asgiref==3.8.0
attrs==25.3.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==25.1
django-phone-field==1.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.4
pytest-django==4.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.4.4
typing_extensions==4.13.1
tzdata==2024.1
uritemplate==4.1.1
